Cleaned code, changed database to replace rows because of limited storage, fixed wrong function call

+++ src/violation_manager/violation_manager.py
@@ -1,5 +1,5 @@
 from src.violation_monitor.violation_monitor import ViolationMonitor, Drone, Pilot
-from src.violation_manager.structs import OnViolationData, OnViolationExpiredData
+from src.violation_manager.structs import ViolationData
 from src.utils.delayed_task import DelayedTask
 from src.database.database import Database
 from typing import Callable
@@ -19,11 +19,11 @@ class ViolationManager:
 
         self.database: Database = database
         self.delete_data_after: int = delete_data_after
-        self.on_expire: Callable[[OnViolationExpiredData], None] | None = None
-        self.on_violation: Callable[[OnViolationData], None] | None = None
+        self.on_expire: Callable[[ViolationData], None] | None = None
+        self.on_violation: Callable[[ViolationData], None] | None = None
         self.violation_monitor: ViolationMonitor | None = None
 
-    def fetch_violations(self) -> list[OnViolationData]:
+    def fetch_violations(self) -> list[ViolationData]:
         """
         Fetches all violations from the database
 
@@ -48,17 +48,15 @@ class ViolationManager:
             ORDER BY timestamp ASC
         """)
 
-        violations: list[OnViolationData] = []
+        violations: list[ViolationData] = []
         rows = result.fetchall()
         for serial_num, x, y, distance, timestamp, passed_time in rows:
             drone = Drone(serial_number=serial_num, position_x=x, position_y=y, distance=distance, timestamp=timestamp)
             drone.pilot = pilots.get(serial_num, default_pilot)
-            violation_data = OnViolationData(drone=drone)
-            violation_data['distance'] = round(distance, 2)
-            violation_data['position_x'] = round(x, 2)
-            violation_data['position_y'] = round(y, 2)
-            violation_data['delete_after'] = max(self.delete_data_after - passed_time, 0)
 
+            delete_after = max(self.delete_data_after - passed_time, 0)
+            violation_data = ViolationData(delete_after=delete_after)
+            violation_data.from_drone(drone)
             violations.append(violation_data)
 
         return violations
@@ -71,15 +69,33 @@ class ViolationManager:
         :return: None
         """
 
+        cursor = self.database.db.execute("""
+            SELECT EXISTS (
+                SELECT 1
+                FROM drones
+                WHERE serial_number = ? AND STRFTIME('%S', 'now') - STRFTIME('%S', timestamp) > ?
+            );
+            """, (drone.serial_number, self.delete_data_after / 60))
+
+        # Handle case where drone violation has been updated since the violation was created
+        has_expired = cursor.fetchone()[0]
+        if not has_expired:
+            return
+
         self.database.db.execute("""
             DELETE FROM drones WHERE serial_number = ?
-        """, (drone.serial_number,))
+            """, (drone.serial_number,))
+
+        self.database.db.execute("""
+            DELETE FROM pilots WHERE drone_serial_number = ?
+            """, (drone.serial_number,))
 
         self.database.db.commit()
 
-        data = OnViolationExpiredData(drone=drone)
+        violation_data = ViolationData()
+        violation_data.from_drone(drone)
         if self.on_expire is not None:
-            self.on_expire(data)
+            self.on_expire(violation_data)
 
     def _on_violation(self, drone: Drone):
         """
@@ -90,13 +106,14 @@ class ViolationManager:
         """
 
         self.database.db.execute("""
-            INSERT INTO drones(serial_number, position_x, position_y, distance, timestamp)
-            VALUES (?, ?, ?, ?, ?)""", (
+            INSERT OR REPLACE INTO drones(serial_number, position_x, position_y, distance, timestamp)
+            VALUES (?, ?, ?, ?, ?)
+            """, (
             drone.serial_number,
             drone.position_x,
             drone.position_y,
             drone.distance,
-            drone.timestamp,
+            drone.timestamp
         ))
 
         if drone.pilot is not None:
@@ -112,17 +129,14 @@ class ViolationManager:
             ))
 
         self.database.db.commit()
+        violation_data = ViolationData(delete_after=self.delete_data_after)
+        violation_data.from_drone(drone)
 
-        data = OnViolationData(drone=drone)
-        data['distance'] = round(drone.distance, 2)
-        data['position_x'] = round(drone.position_x, 2)
-        data['position_y'] = round(drone.position_y, 2)
-        data['delete_after'] = self.delete_data_after
         if self.on_violation is not None:
-            self.on_violation(data)
+            self.on_violation(violation_data)
 
         if self.delete_data_after > 0:
-            delayed_task = DelayedTask(task=self.on_expire, delay=self.delete_data_after, drone=drone)
+            delayed_task = DelayedTask(task=self._on_expire, delay=self.delete_data_after, drone=drone)
             delayed_task.start()
 
     def create_monitor(self, update_interval: int, ndz_origin: tuple[float, float], ndz_radius: int) -> None:
